# docs/swagger.yaml
definitions:
...
  response.response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      time:
        type: string
    type: object
  userService.DeductMarginReq:
    properties:
      amount:
        description: 扣除金额
        type: integer
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.DeductMarginResp:
    type: object
  userService.GetMarginResp:
    properties:
      amount:
        description: 保证金数量
        type: integer
    type: object
  userService.KycReq:
    properties:
      name:
        description: 姓名
        type: string
      phone:
        description: 手机号
        type: string
      roleCode:
        description: 申请成为的角色
        type: integer
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.KycResp:
    type: object
  userService.LoginReq:
    properties:
      authCode:
        description: 验证码
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  userService.LoginResp:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/userService.User'
    type: object
  userService.PayMarginReq:
    properties:
      amount:
        description: 缴纳金额
        type: integer
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.PayMarginResp:
    type: object
  userService.RegisterUserReq:
    properties:
      displayName:
        description: 昵称，长度6-20
        type: string
      password:
        description: 密码，长度6-20
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  userService.RegisterUserResp:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  userService.Role:
    properties:
      code:
        description: 角色代码
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        description: 角色名
        type: string
      updatedAt:
        type: string
    type: object
  userService.UpdateUserStatusReq:
    properties:
      adminId:
        description: 管理员uuid
        type: string
      status:
        description: 状态:1=拉黑，2=解除拉黑
        type: integer
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.UpdateUserStatusResp:
    type: object
  userService.User:
    properties:
      avatar:
        description: 头像图片地址
        type: string
      ban:
        description: 用户是否被拉黑
        type: boolean
      createdAt:
        type: string
      deletedAt:
        type: string
      displayName:
        description: 昵称
        type: string
      id:
        type: integer
      margin:
        description: 保证金
        type: integer
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      role:
        allOf:
        - $ref: '#/definitions/userService.Role'
        description: 角色
      updatedAt:
        type: string
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.WithdrawMarginReq:
    properties:
      amount:
        description: 提现金额
        type: integer
      userId:
        description: 用户uuid
        type: string
    type: object
  userService.WithdrawMarginResp:
    type: object
info:
  contact: {}
paths:
...
  /user:
    post:
      consumes:
      - application/json
      description: 注册用户
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.RegisterUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.RegisterUserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 注册用户
      tags:
      - user模块
  /user/authcode:
    get:
      consumes:
      - application/json
      description: 获取验证码
      parameters:
      - description: 无参数
        in: query
        name: None
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.RegisterUserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 获取验证码
      tags:
      - user模块
  /user/kyc:
    post:
      consumes:
      - application/json
      description: Kyc
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.KycReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.KycResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: Kyc
      tags:
      - user模块
  /user/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.LoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 登录
      tags:
      - user模块
  /user/margin:
    delete:
      consumes:
      - application/json
      description: 管理员扣除保证金
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.DeductMarginReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.DeductMarginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 扣除保证金
      tags:
      - user模块
    get:
      consumes:
      - application/json
      description: 查询保证金
      parameters:
      - description: 用户uuid
        in: query
        name: userId
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.GetMarginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 查询保证金
      tags:
      - user模块
    post:
      consumes:
      - application/json
      description: 乙方缴纳保证金
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.PayMarginReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.PayMarginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 缴纳保证金
      tags:
      - user模块
    put:
      consumes:
      - application/json
      description: 乙方保证金提现
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.WithdrawMarginReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.WithdrawMarginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 保证金提现
      tags:
      - user模块
  /user/status:
    put:
      consumes:
      - application/json
      description: 管理员拉黑、解除拉黑用户
      parameters:
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/userService.UpdateUserStatusReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userService.UpdateUserStatusResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.response'
      summary: 拉黑、解除拉黑用户
      tags:
      - user模块
swagger: "2.0"